+++ src/main/java/ru/vyarus/guice/persist/orient/db/pool/object/OObjectDatabaseTxFixed.java
@@ -3,6 +3,8 @@ package ru.vyarus.guice.persist.orient.db.pool.object;
 import com.orientechnologies.orient.core.command.OCommandRequest;
 import com.orientechnologies.orient.core.db.ODatabaseDocumentInternal;
 import com.orientechnologies.orient.object.db.OObjectDatabaseTx;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import edu.umd.cs.findbugs.annotations.SuppressWarnings;
 
 /**
  * Fixed version of {@link OObjectDatabaseTx}: corrects some cases of result conversions (bugs which will not be fixed
@@ -13,24 +15,27 @@ import com.orientechnologies.orient.object.db.OObjectDatabaseTx;
  * @author Vyacheslav Rusakov
  * @since 28.06.2021
  */
+@SuppressFBWarnings("HE_INHERITS_EQUALS_USE_HASHCODE")
 public class OObjectDatabaseTxFixed extends OObjectDatabaseTx {
 
-    public OObjectDatabaseTxFixed(String iURL) {
+    public OObjectDatabaseTxFixed(final String iURL) {
         super(iURL);
     }
 
-    public OObjectDatabaseTxFixed(ODatabaseDocumentInternal iDatabase) {
+    public OObjectDatabaseTxFixed(final ODatabaseDocumentInternal iDatabase) {
         super(iDatabase);
     }
 
     @Override
-    public <RET extends OCommandRequest> RET command(OCommandRequest iCommand) {
+    @SuppressWarnings("unchecked")
+    public <RET extends OCommandRequest> RET command(final OCommandRequest iCommand) {
         return (RET) new OCommandSQLPojoWrapperFixed(this, underlying.command(iCommand));
     }
 
-    // fixing access required for fixed wrapper command
     @Override
-    public void convertParameters(Object... iArgs) {
+    // changing access, required for fixed wrapper command
+    @SuppressWarnings("PMD.UselessOverridingMethod")
+    public void convertParameters(final Object... iArgs) {
         super.convertParameters(iArgs);
     }
 }